Move resolvers and schema to graphql folder
