+++ starters/deno-oak-denodb/src/graphql/interfaces/graphql_interfaces.ts
@@ -1,0 +1,5 @@
+export interface TechnologyInput {
+  displayName: string;
+  description: string;
+  url: string;
+}
