+++ starters/deno-oak-denodb/src/graphql/resolvers/queryHandler.ts
@@ -1,11 +1,11 @@
-import { Technologies } from '../model/technology.ts';
+import { Technologies } from '../../db/model/technology.ts';
 
 export const getTechnologies = async () => {
 	const technologies = await Technologies.all();
 	return technologies;
 };
 
-export const getTechnology = async (_: unknown, { id }: any) => {
+export const getTechnology = async (_: unknown, { id }: { id: string }) => {
 	const technology = await Technologies.find(id);
 	return technology;
 };
